update handleFileInput
HTHL에서도 input type을 이미지로 제한했지만, 코드상으로도 한 번 더 검증하는 로직 및 입력 최대 사이즈 설정

+++ renew/src/hooks/useCardMaker.js
@@ -5,6 +5,8 @@ import { deleteCloudinaryImage, uploadCloudinaryImage } from '../api'
 import { useDispatch } from 'react-redux'
 import { createCard } from '../store/cardsSlice'
 
+const MAX_PROFILE_SIZE = 3000000 // 3MB
+
 export default function useCardMaker() {
     const { state, dispatch } = useContext(CardContext)
     const dispatchRedux = useDispatch()
@@ -32,6 +34,16 @@ export default function useCardMaker() {
 
     const handleFileInput = async (e) => {
         setFileLoading(true)
+
+        if (
+            e.target.files[0]?.size > MAX_PROFILE_SIZE ||
+            !e.target.files[0]?.type.startsWith('image')
+        ) {
+            setFileLoading(false)
+            // TODO: show popup in Maker Component
+            return
+        }
+
         const data = await uploadCloudinaryImage(e.target.files[0])
         if (state.profile.url)
             deleteCloudinaryImage(
